fix: isolate Layout component exports for Fast Refresh (#6993)
