+++ client/src/lib/index.js
@@ -216,3 +216,5 @@ export * from './riggingReasons.js';
 export * from './usdzExport.js';
 export * from './eidoverseFrame.js';
 export * from './songBookDraft.js';
+export * from './layoutRoutes.js';
+export * from './navPresentation.js';
