+++ client/src/lib/navPresentation.js
@@ -1,0 +1,292 @@
+import {
+  Wrench,
+  Home,
+  Package,
+  FileText,
+  Bot,
+  History,
+  Code2,
+  Activity,
+  BarChart3,
+  Cpu,
+  Gauge,
+  FlaskConical,
+  Braces,
+  Play,
+  ScrollText,
+  Camera,
+  Brain,
+  Heart,
+  Fingerprint,
+  CheckCircle,
+  Dna,
+  Download,
+  Film,
+  MessageSquare,
+  Palette,
+  PenLine,
+  Sparkles,
+  Target,
+  Clock,
+  Calendar,
+  CalendarDays,
+  GraduationCap,
+  Settings,
+  Users,
+  CalendarClock,
+  Upload,
+  SquareTerminal,
+  Globe,
+  Newspaper,
+  Orbit,
+  Ticket,
+  Network,
+  Flame,
+  BarChart2,
+  Cigarette,
+  HeartPulse,
+  ClipboardList,
+  ListChecks,
+  Compass,
+  Feather,
+  Scale,
+  Lightbulb,
+  GitBranch,
+  Link2,
+  ListMusic,
+  Database,
+  Shield,
+  ShieldCheck,
+  KeyRound,
+  Lock,
+  Wand2,
+  Rocket,
+  Zap,
+  Inbox,
+  RefreshCw,
+  Dog,
+  FileInput,
+  FilePen,
+  MessageCircle,
+  Radio,
+  TrendingUp,
+  Swords,
+  HardDrive,
+  Layers,
+  MessagesSquare,
+  BookOpen,
+  NotebookPen,
+  Mic,
+  Rss,
+  Archive,
+  Eraser,
+  Share2,
+  Music,
+  Workflow as WorkflowIcon,
+  ChartGantt,
+  Clapperboard,
+  PersonStanding,
+  Box,
+  Blocks,
+  Boxes,
+  Gamepad2,
+  Waypoints,
+  AtSign,
+  Drama,
+  UserRound,
+  Video,
+} from 'lucide-react';
+
+// NAV_COMMANDS owns every structural field shared with the sidebar. This map
+// intentionally contains presentation only: giving a manifest path an icon is
+// what opts that destination into the sidebar. External and runtime-hydrated
+// rows have no manifest destination and stay explicitly local below.
+export const NAV_PRESENTATION = {
+  '/': { icon: Home, single: true },
+  '/review': { icon: ClipboardList, single: true },
+  '/eidoverse': { icon: Orbit, single: true },
+  '/apps': { icon: Package, dynamic: 'apps' },
+  '/brain/config': { icon: Settings },
+  '/brain/daily-log': { icon: NotebookPen },
+  '/brain/digest': { icon: Calendar },
+  '/brain/feeds': { icon: Rss },
+  '/brain/graph': { icon: Network },
+  '/brain/ideas': { icon: Lightbulb },
+  '/brain/import': { icon: Upload },
+  '/brain/inbox': { icon: MessageSquare },
+  '/brain/links': { icon: Link2 },
+  '/brain/memory': { icon: Database },
+  '/brain/notes': { icon: FileText },
+  '/brain/spotify': { icon: Music },
+  '/brain/youtube': { icon: Video },
+  '/rapid-reader': { icon: Zap },
+  '/songbook': { icon: ListMusic },
+  '/timeline': { icon: CalendarClock },
+  '/tribe': { icon: Users },
+  '/brain/trust': { icon: Shield },
+  '/wiki/overview': { icon: BookOpen },
+  '/calendar/agenda': { icon: CalendarDays },
+  '/calendar/config': { icon: Settings },
+  '/calendar/day': { icon: Calendar },
+  '/calendar/lifetime': { icon: Clock },
+  '/calendar/month': { icon: CalendarDays },
+  '/calendar/review': { icon: ClipboardList },
+  '/calendar/sync': { icon: RefreshCw },
+  '/calendar/week': { icon: CalendarDays },
+  '/cos/agents': { icon: Cpu },
+  '/cos/briefing': { icon: Newspaper },
+  '/cos/config': { icon: Settings },
+  '/cos/digest': { icon: Calendar },
+  '/feature-agents': { icon: Wand2 },
+  '/cos/gsd': { icon: Compass },
+  '/cos/health': { icon: Activity },
+  '/cos/learning': { icon: GraduationCap },
+  '/cos/memory': { icon: Brain },
+  '/cos/mind': { icon: MessageSquare },
+  '/cos/run-events': { icon: ScrollText },
+  '/cos/runs': { icon: Play },
+  '/cos/schedule': { icon: Clock },
+  '/agents': { icon: Users },
+  '/cos/productivity': { icon: BarChart2 },
+  '/cos/jobs': { icon: Bot },
+  '/cos/tasks': { icon: FileText },
+  '/cos/workflow': { icon: ChartGantt },
+  '/messages/beeper': { icon: MessageCircle },
+  '/messages/config': { icon: Settings },
+  '/messages/contacts': { icon: Users },
+  '/messages/drafts': { icon: FilePen },
+  '/messages/imessage': { icon: MessageSquare },
+  '/messages/inbox': { icon: Inbox },
+  '/messages/signal': { icon: MessageSquare },
+  '/stacker-news': { icon: Newspaper },
+  '/messages/sync': { icon: RefreshCw },
+  '/x': { icon: AtSign },
+  '/3d': { icon: Boxes },
+  '/authors': { icon: FilePen },
+  '/catalog': { icon: Sparkles },
+  '/creative-commission': { icon: CalendarClock },
+  '/creative-director': { icon: Clapperboard },
+  '/video': { icon: Clapperboard },
+  '/pipeline/editorial-checks': { icon: ListChecks },
+  '/fableloom': { icon: Waypoints },
+  '/game': { icon: Gamepad2 },
+  '/importer': { icon: FileInput },
+  '/media': { icon: Layers },
+  '/mood-boards': { icon: Palette },
+  '/music': { icon: Mic },
+  '/music-video': { icon: Music },
+  '/rounds': { icon: Music },
+  '/pipeline': { icon: WorkflowIcon, dynamic: 'pipelineSeries' },
+  '/sharing': { icon: Share2 },
+  '/sprites': { icon: PersonStanding },
+  '/start-story': { icon: Rocket },
+  '/story-builder': { icon: Wand2 },
+  '/media/threejs': { icon: Box },
+  '/universes': { icon: Globe, dynamic: 'universes' },
+  '/writers-room': { icon: NotebookPen },
+  '/devtools/agents': { icon: Cpu },
+  '/ambient': { icon: Sparkles },
+  '/browser': { icon: Globe },
+  '/capabilities': { icon: Compass },
+  '/devtools/runner': { icon: Code2 },
+  '/data': { icon: HardDrive },
+  '/devtools/datadog': { icon: Dog },
+  '/devtools/flows': { icon: WorkflowIcon },
+  '/devtools/github': { icon: GitBranch },
+  '/devtools/history': { icon: History },
+  '/devtools/image-clean': { icon: Eraser },
+  '/instances': { icon: Network },
+  '/devtools/jira': { icon: Ticket },
+  '/devtools/jira/reports': { icon: FileText },
+  '/loops': { icon: RefreshCw },
+  '/devtools/processes': { icon: Activity },
+  '/devtools/quota-burn': { icon: Flame },
+  '/security': { icon: Camera },
+  '/shell': { icon: SquareTerminal },
+  '/system-resources': { icon: Activity },
+  '/uploads': { icon: Upload },
+  '/devtools/usage': { icon: BarChart3 },
+  '/devtools/video-download': { icon: Film },
+  '/workspace-contexts': { icon: Layers },
+  '/goals/list': { icon: Target, single: true },
+  '/meatspace/age': { icon: Clock },
+  '/meatspace/alcohol': { icon: Activity },
+  '/meatspace/blood': { icon: HeartPulse },
+  '/meatspace/body': { icon: Scale },
+  '/meatspace/health': { icon: Heart },
+  '/meatspace/export': { icon: FileText },
+  '/meatspace/genome': { icon: Dna },
+  '/meatspace/lifestyle': { icon: ClipboardList },
+  '/meatspace/nicotine': { icon: Cigarette },
+  '/meatspace/overview': { icon: Activity },
+  '/meatspace/settings': { icon: Settings },
+  '/models/3d': { icon: Boxes },
+  '/models/embeddings': { icon: Braces },
+  '/models/harnesses': { icon: Blocks },
+  '/models/llms': { icon: Cpu },
+  '/models/loras': { icon: Sparkles },
+  '/models/media': { icon: HardDrive },
+  '/models/comparison': { icon: Gauge },
+  '/models/performance': { icon: Gauge },
+  '/local-llm/playground': { icon: FlaskConical },
+  '/models/status': { icon: Activity },
+  '/models/training': { icon: GraduationCap },
+  '/settings/ai-assignments': { icon: Bot },
+  '/settings/api-access': { icon: Globe },
+  '/api-reference/catalog': { icon: Braces },
+  '/settings/autofixer': { icon: Wrench },
+  '/settings/backup': { icon: Download },
+  '/settings/credentials': { icon: KeyRound },
+  '/models/code-reviewers': { icon: ShieldCheck },
+  '/settings/database': { icon: Database },
+  '/settings/features': { icon: ListChecks },
+  '/settings/general': { icon: Settings },
+  '/settings/mortalloom': { icon: Activity },
+  '/openclaw': { icon: MessagesSquare },
+  '/settings/orchestration': { icon: Cpu },
+  '/prompts': { icon: FileText },
+  '/ai': { icon: Bot },
+  '/settings/security': { icon: Lock },
+  '/settings/sharing': { icon: Share2 },
+  '/settings/telegram': { icon: MessageSquare },
+  '/settings/voice': { icon: Mic },
+  '/digital-twin/accounts': { icon: Globe },
+  '/digital-twin/appearance': { icon: Camera },
+  '/ask': { icon: MessageCircle },
+  '/digital-twin/autobiography': { icon: PenLine },
+  '/digital-twin/avatar-bio': { icon: UserRound },
+  '/character': { icon: Swords },
+  '/digital-twin/documents': { icon: FileText },
+  '/digital-twin/enrich': { icon: Sparkles },
+  '/digital-twin/export': { icon: Download },
+  '/digital-twin/goals': { icon: Target },
+  '/digital-twin/identity': { icon: Fingerprint },
+  '/digital-twin/import': { icon: Upload },
+  '/insights/overview': { icon: Lightbulb },
+  '/digital-twin/interview': { icon: MessageSquare },
+  '/digital-twin/legacy': { icon: Package },
+  '/digital-twin/overview': { icon: Heart },
+  '/digital-twin/personality': { icon: Brain },
+  '/digital-twin/personas': { icon: Drama },
+  '/privacy/overview': { icon: Shield },
+  '/digital-twin/taste': { icon: Palette },
+  '/digital-twin/test': { icon: CheckCircle },
+  '/digital-twin/time-capsule': { icon: Archive },
+  '/digital-twin/voice': { icon: Mic },
+  '/post/config': { icon: Settings },
+  '/post/explore': { icon: Compass },
+  '/post/history': { icon: History },
+  '/post/launcher': { icon: Play },
+  '/post/memory': { icon: Brain },
+  '/post/morse': { icon: Radio },
+  '/post/plan': { icon: ListChecks },
+  '/post/progress': { icon: TrendingUp },
+  '/post/rhetoric': { icon: Feather },
+  '/post/wordplay': { icon: MessageCircle },
+};
+
+export const SECTIONS_BEFORE_GOALS = [
+  'Brain', 'Calendar', 'Chief of Staff', 'Comms', 'Create', 'Dev Tools',
+];
+export const SECTIONS_AFTER_GOALS = ['Health', 'Models', 'Settings'];
+export const SECTIONS_BELOW_MORE = ['Identity', 'POST'];
